feat: refactor the icon component
The Icon component when was used, the component didn't have any options to change the color of the svg, so if you have a arrow who is black in a button and the same arrow is used in another button, we couldn't change the color.

This commit creates a new icon component, changing the file sctructure and using a useMemo to not reender the same icon if he use in the next render.